requirements.txt
openai>=1.0.0
PyPDF2
python-docx
python-dotenv